requirements.txt
numpy
scipy
torch
torchvision
detectron2
opencv-contrib-python
imageio
imageio-ffmpeg
crefi
pycocotools
Pillow
click
matplotlib
scikit-image